Bug 1710 - The Incredible Shrinking ToolBar
The worst symptoms of this are fully fixed.
1: The device info now uses all the space in the toolbar, no matter how the toolbar is resized.
2: The undocked toolbar can now be moved around repeatedly without changing size at all.

There is a small residual - a much toolbar size change of a few pixels on docking.  The contents though still fit nicely.  The 'incredible' part of this bug is gone.  I'd like us to close this bug as the residual of a few pixels change in the toolbar size on repeated docking and undocking scarcely matters.