Change TrackPrefs Pinned/Unpinned display.
The pin and its position can be set directly from the ruler, which is always visible.  Displaying the pin/unpinned state and pin position is probably a bit like displaying the toolbar visibility states and positions in prefs - overload of settings that are easier set another way.  I've hidden them for now.  We should probably come up with a policy as to what prefs should be shown in prefs and what should not.