Use the TimeQueue to communicate all play head position updates...
 ... Scrub is no longer a special case.

The PortAudio callback thread only consumes the values without calculating
them.  The AudioThread calculates all now, and will have freedom to
calculate differently.

Because of the time queue granularity, this requires AudioIO::FillBuffers to
produce a little bit of extra silence, when not looping or scrubbing, so that
the consumer thread will detect its termination condition.