Refactor Normalize in preparation of loudness gating.
This commit switches the processing order of stereo pairs from
one track after another to both stereo tracks simultaneously.
In order to do loudness gating of a stereo pair with a fixed
amount of memory it is necessary to have the same loudness
block of both stereo channels available at the same time.