Bug 1749 - Timer record dialog: wxWidgets assert failure
The problem was that the end time was set to prohibit changes that brought it
before the start time.  However an update to start time could update the end time
before the end time updated its legal range.

The prohibition on going backwards in time is only for user interaction with
that control, so we now clear the legal range before update and then recreate it.

This change also fixes:
Bug 1978 - Windows: Timer Record - a "debugging check" dialog is shown to the user

In passing I also noticed that the controls were being updated every 50ms.
This is totally pointless as they only show the nearest second.  So I set the granularity
for control updates to 1s. (1000ms).