Bug 2009: Windows+Linux: Intermittent truncation of the end of a recording
Problem: up to about 300ms was intermittently truncated from the end of recording.
The bug was introduced by commit a0256e9.

The fix: in this commit, the line:
the line:
mRecordingSchedule = {}; // free arrays
was added to the lambda in the definition of cleanup at the start of AudioIO::StopStream()

If this line is changed to:
mRecordingSchedule.mCrossfadeData.clear();  // free arrays
then this has been found to fix the bug, and the vector is still cleared.

Clearly, the reinitialization of one or more of the other data members of mRecordingSchedule was causing the bug.
Which of the these reinitializations are the problem, and why is not known, and due to the nature of the problem could well take some time to find out.