Apply macro sub menu: ensure first letter selection
Problem: Palette has the access key P, so any macro with name beginning with P cannot be selected by pressing P, since access keys have precedence over first letters.

Fix: remove access key, which is not needed since Palette is the first item, and so is selected when the sub menu is opened.