Fix failure of Nyquist to translate some multiline strings...
... Don't trim leading whitespace from source file lines before passing them
to eval.

So when a string in Lisp code is (_ "like
   this")

(the Lisp reader allows embedded newlines in quoted strings),

then the string extracted to the .pot file by xgettext contains the spaces, and
the string literal seen by our Lisp interpreter will also retain the spaces and
match the msgid in the catalog exactly.