Bug 2025: Windows: Wasapi: seek commands can cause crash
Problem: With host set to WASAPI: if the playback position is less that about ten seconds before the end of any of the tracks and the user presses one of the left/arrow keys to seek, then a bug alert appears, and it's seems like it's impossible to recover from this.

libsoxr 0.1.3, first used in Audacity 2.3.0, crashes with constant rate resampling if you try to reuse the resampler after it has flushed.

Fix: work around which is the same as used for bug 1887 - recreate resamplers, rather than reusing them.