Bug 2028 - Mac: Audacity crashes when right clicking in label after copying label text to another app
With the steps in bug 2028, FindFocus() could return a nullPtr - which is now safe.
No right context menu will appear though.

This looks like a bug in wxWidgets.

We could take more/different steps to find the parent for the LabelTrack, if we really want to do that.  However, this change is enough to close the bug.