Bug 2079: FreeBSD compilation broken by -latomic use.
To cut a long story short the least painful way to fix this is to link
against libatomic if it is present and to not otherwise.

clang needs libatomic if either:
  1. it's using libstdc++ during build;
  2. it's using libc++ but libcomiler-rt has been built without support of
     atomic functions.

It seems quite a job to find out how clang has been set up. Therefore
I'm prorosing this simpler patch which gives the desired result both on
FreeBSD and Archlinux.

Also I've updated AX_COMPILER_VENDOR macro. Now it supports clang and
thus detecting the type of the compiler can be expressed more elegantly.