Remove unused overloads of ShuttleGuiBase::TieNumericTextBox...
... those taking wxString as default value, or reference to wxString as the
target variable.

There remain one that takes reference to int as target, and one taking
reference to double; and one taking a config path and a double default value,
passed by const reference.  Some of the calls to the last pass an int default
value which was, and still is, implicitly converted to double.