Consistent cap. of config paths in EffectsPrefs with family names...
... That is:  EffectsPrefs.cpp is consistent with overrides of
EffectDefinitionInterface::GetFamily for each of the six (separately).

(The six aren't consistent with each other.  LADSPA really is an acronym, but
Vamp is not.)

See VAMPEFFECTS_FAMILY, which isn't up-cased, and LADSPAEFFECTS_FAMILY,
which is.  We change EffectsPrefs.cpp to match those constants.

See also where PluginManager.cpp uses
PluginDescriptor::GetEffectFamily, which is where the preferences written by
EffectsPrefs.cpp are used.

Does this break compatibility of old config files?  No, because lookup of
keys in wxConfigFile is case-insensitive, as determined by the preprocessor
constant wxCONFIG_CASE_SENSITIVE, which is false for all of our platforms.

So, no real effects here, but the consistency of string literals is reassuring.

It would be better, though, not to rely on such a coincidence of literals.