Revert "Bug1848: Mac VoiceOver should read names of choice controls..."
 This reverts commit d2c112600d38f0894fe783f8dfd8861b5e604fce.

On Windows, it is intentional that SetAccessible is not called for wxChoice in ShuttleGui.
See the commit message of f9ee3cd for background.

Changing this during code freeze and close to creating release candidates would have the danger of introducing bugs.

Fixing this bug needs further discussion. Hopefully the Mac accessibility code can be changed to bring it into line with the changes in the wxWidets accessibility code when we moved to 3.1.1