Don't destroy auto-save file written by different architecture...
... the error checking might not be complete, but it is sufficient for the
observed cases, where switching between 32 and 64 bit Mac builds causes
auto-recovery in one build to destroy the data saved by the other build.

Now instead, you will see an error message, recommending that you run the
same version of Audacity that produced the file.

Note that decoding of autosave files can also (less commonly) happen with
a command-line argument, and a message is written to standard out.  Give the
same message in that case.

Localization of this changed message unfortunately can't happen this late in
2.3.2 development.