New script graph.pl to analyze dependencies...
... To use it, first install graphViz.  Then change to the scripts directory
and run ./graph.pl .  Then view the resulting image file in a browser.

See an acyclic graph, each node labeled with the list of .cpp files in
one strongly connected component of the linkage dependencies as determined
by scanning the #include directives in the sources.

It worked for me in macOS.