Bug 2178 - Export to (external program) fails with errors [WINDOWS]
This fixes the problem on Windows that if a user does not specify a suffix then
ffmpeg does not know what format to use and gives a cryptic error message,
if exporting using

ffmpeg -i - "filename"

Now, if no suffix is given, .wav is appended in the command sent to ffmpeg.
If a user gives a suffix, e.g. 'myfile.wma' then no suffix is added.