Bug 1548 - Newly recorded tracks do not acquire focus
Problem: The previous fix (+ possibly later changes) meant that this was working for mono tracks but not stereo. When recording in a new stereo track, the first track in the project became the focus.
The existing code sets the second of the two new tracks to be the focus before the tracks have been grouped into a stereo pair.

Fix: Set the first of the new tracks to be the focus.