Make Graphic EQ vertically stretchable
Drag to enlarge the window, and the length of the graphic sliders.
This was part of splitting EQ.  The split effect would look strange if we did not do this.

Graphic EQ sliders are now in a sizer, and so we no longer lay them out 'by hand' in the way we previously did.